CREATE modular OID4VP wallet and SSF stream modules
