+++ frontend/src/pages/SSFSandbox.tsx
@@ -7,367 +7,35 @@
  * Each browser session gets isolated data via session ID.
  */
 
-import { useState, useEffect, useCallback, useMemo, useRef } from 'react'
-import { motion, AnimatePresence } from 'framer-motion'
+import { useState, useEffect, useCallback, useMemo, type ElementType } from 'react'
+import { motion } from 'framer-motion'
 import { 
   Radio, ChevronRight, Shield, AlertTriangle, 
-  User, CheckCircle, Clock, Zap, 
-  RotateCcw, Send, ArrowDownLeft, Bell, 
-  Lock, UserX, UserCheck, Terminal,
-  Sparkles, Play, Key, Book, Info,
-  ChevronDown, Code, Copy, Check, Square,
-  XCircle, Globe
+  RotateCcw, Lock, UserX, UserCheck, Terminal,
+  Sparkles, Play, Key, Info, Square, Code, User, ChevronDown,
+  XCircle
 } from 'lucide-react'
-import { TokenInspector } from '../components/lookingglass/TokenInspector'
+import { TokenInspector } from '../lookingglass/components/inspectors/TokenInspector'
 import { SSFSandboxSEO } from '../components/common/SEO'
+import {
+  StatusBadge as SharedStatusBadge,
+  type StatusBadgeVariant,
+} from '../lookingglass/components/shared'
+import { getOrCreateSSFSessionId, ssfFetch, useSSFEventStream } from '../hooks/useSSFEventStream'
+import type { DecodedSET, EventDef, SecurityState, Subject, SSEPipelineEvent } from '../lookingglass/ssf/types'
+import { SSFFlowPanel } from '../lookingglass/components/ssf/SSFFlowPanel'
 
-// ============================================================================
-// Session Management - Isolates each user's sandbox
-// ============================================================================
-
-function getOrCreateSessionId(): string {
-  const STORAGE_KEY = 'ssf_session_id'
-  let sessionId = localStorage.getItem(STORAGE_KEY)
-  
-  if (!sessionId) {
-    // Generate a random session ID
-    sessionId = 'sess_' + Math.random().toString(36).substring(2, 15) + 
-                Math.random().toString(36).substring(2, 15)
-    localStorage.setItem(STORAGE_KEY, sessionId)
-  }
-  
-  return sessionId
+const SSF_STATUS_BADGE_VARIANTS: Record<string, StatusBadgeVariant> = {
+  completed: { bg: 'bg-green-500/10', border: 'border-green-500/30', text: 'text-green-400', label: 'Completed', shortLabel: 'Done' },
+  executing: { bg: 'bg-amber-500/10', border: 'border-amber-500/30', text: 'text-amber-400', label: 'Executing...', shortLabel: 'Running' },
+  error: { bg: 'bg-red-500/10', border: 'border-red-500/30', text: 'text-red-400', label: 'Error', shortLabel: 'Error' },
 }
-
-// Create fetch wrapper that includes session ID
-function ssfFetch(url: string, options: RequestInit = {}): Promise<Response> {
-  const sessionId = getOrCreateSessionId()
-  
-  return fetch(url, {
-    ...options,
-    headers: {
-      ...options.headers,
-      'X-SSF-Session': sessionId,
-    },
-  })
-}
-
-// ============================================================================
-// Types
-// ============================================================================
-
-interface Subject {
-  id: string
-  stream_id: string
-  format: string
-  identifier: string
-  display_name: string
-  status: string
-  active_sessions: number
-  last_activity: string | null
-  created_at: string
-}
-
-interface SecurityState {
-  email: string
-  sessions_active: number
-  account_enabled: boolean
-  password_reset_required: boolean
-  tokens_valid: boolean
-  last_modified: string
-  modified_by: string
-}
-
-interface FlowEvent {
-  id: string
-  type: 'info' | 'request' | 'response' | 'token' | 'crypto' | 'security' | 'action' | 'error'
-  title: string
-  description: string
-  timestamp: Date
-  rfcReference?: string
-  data?: Record<string, unknown>
-}
-
-interface DecodedSET {
-  jti: string
-  iss: string
-  aud: string[]
-  iat: string
-  sub_id: { format: string; email?: string }
-  events: Array<{
-    type: string
-    metadata: { name: string; category: string; response_actions: string[]; zero_trust_impact: string }
-    payload: Record<string, unknown>
-  }>
-  header: Record<string, unknown>
-  raw_token: string
-}
-
-// ============================================================================
-// SSE Pipeline Types
-// ============================================================================
-
-interface SSEPipelineEvent {
-  source: 'transmitter' | 'receiver'
-  event: {
-    type: string
-    timestamp: string
-    event_id: string
-    session_id?: string
-    subject_id?: string
-    event_type?: string
-    data: Record<string, unknown>
-  }
-}
-
-interface CapturedHTTPExchange {
-  label: string
-  request: { method: string; url: string; status_code?: number; headers: Record<string, string>; body?: string }
-  response: { method?: string; url?: string; status_code: number; headers: Record<string, string>; body?: string }
-  duration_ms: number
-  timestamp: string
-  session_id?: string
-}
-
-// SSE event type -> FlowEvent mapping.
-// Only events that carry real, observable data are shown.
-// Metadata-only announcements (event_queued, delivery_started, delivery_success)
-// are dropped — the captured http_exchange IS the delivery with real headers/body.
-function mapSSEToFlowEvent(sse: SSEPipelineEvent): FlowEvent | null {
-  const { source, event } = sse
-  const base = {
-    id: crypto.randomUUID(),
-    timestamp: new Date(event.timestamp),
-    data: event.data,
-  }
-
-  // ── Transmitter events ──────────────────────────────────────────────
-  if (source === 'transmitter') {
-    switch (event.type) {
-      case 'action_triggered': {
-        const meta = event.data?.metadata as Record<string, unknown> | undefined
-        return {
-          ...base, type: 'info' as const,
-          title: `Transmitter: ${(meta?.name as string) || 'Event Triggered'}`,
-          description: `Subject: ${event.subject_id || '?'}  |  ${(meta?.category as string) || 'SSF'}`,
-          rfcReference: 'SSF §4',
-        }
-      }
-      case 'set_generated': {
-        const claims = event.data?.claims as Record<string, unknown> | undefined
-        return {
-          ...base, type: 'crypto' as const,
-          title: 'Transmitter: SET Created',
-          description: `JTI: ${(claims?.id as string)?.slice(0, 8) || '?'}…  |  iss: ${(claims?.issuer as string) || '?'}`,
-          rfcReference: 'RFC 8417 §2',
-        }
-      }
-      case 'set_signed':
-        return {
-          ...base, type: 'crypto' as const,
-          title: 'Transmitter: SET Signed (RS256)',
-          description: `${((event.data?.token as string) || '').length} byte JWT`,
-          rfcReference: 'RFC 7515',
-        }
-
-      // HTTP exchanges are routed to the dedicated Traffic panel
-      case 'http_exchange':
-        return null
-
-      // event_queued, delivery_started, delivery_success are metadata-only
-      // announcements. The real delivery data (headers, body, timing) is in
-      // the http_exchange event shown in the Traffic tab. Showing these here
-      // would display config data (endpoint, retries) that looks like real
-      // traffic but isn't.
-      case 'event_queued':
-      case 'delivery_started':
-      case 'delivery_success':
-        return null
-
-      case 'delivery_failed':
-        return { ...base, type: 'error' as const, title: 'Transmitter: Delivery Failed', description: (event.data?.error as string) || 'Delivery error' }
-
-      default:
-        return null
-    }
-  }
-
-  // ── Receiver events ─────────────────────────────────────────────────
-  if (source === 'receiver') {
-    switch (event.type) {
-      // HTTP exchanges are routed to the dedicated Traffic panel
-      case 'http_exchange':
-        return null
-
-      case 'event_received':
-        return {
-          ...base, type: 'request' as const,
-          title: 'Receiver: SET Received',
-          description: `${(event.data?.token_length as number) || '?'} bytes via ${(event.data?.delivery_method as string) || 'push'}`,
-          rfcReference: 'RFC 8935 §2',
-        }
-      case 'event_verified':
-        return {
-          ...base, type: 'crypto' as const,
-          title: 'Receiver: Signature Verified',
-          description: `${(event.data?.algorithm as string) || 'RS256'} with key ${(event.data?.key_id as string) || '?'}`,
-          rfcReference: 'RFC 7515 §5.2',
-        }
-      case 'event_verify_failed':
-        return { ...base, type: 'error' as const, title: 'Receiver: Verification Failed', description: (event.data?.error as string) || 'Signature verification failed' }
-      case 'event_processed':
-        return {
-          ...base, type: 'security' as const,
-          title: 'Receiver: SET Processed',
-          description: `Completed in ${(event.data?.processing_time_ms as number) || 0}ms`,
-        }
-      case 'response_action':
-        return {
-          ...base, type: 'action' as const,
-          title: `Action: ${(event.data?.action as string) || 'Response'}`,
-          description: `${(event.data?.event_type as string) || ''} (${(event.data?.category as string) || 'SSF'})`,
-        }
-
-      case 'event_processing':
-        return {
-          ...base, type: 'info' as const,
-          title: 'Receiver: Processing SET',
-          description: `Decoding ${(event.data?.event_count as number) || 1} event(s)`,
-          rfcReference: 'RFC 8417 §2',
-        }
-
-      default:
-        return null
-    }
-  }
-
-  return null
-}
-
-// ============================================================================
-// SSE Hook
-// ============================================================================
-
-function useSSFEventStream(sessionId: string) {
-  const [pipelineEvents, setPipelineEvents] = useState<FlowEvent[]>([])
-  const [httpExchanges, setHttpExchanges] = useState<CapturedHTTPExchange[]>([])
-  const [isConnected, setIsConnected] = useState(false)
-  const eventSourceRef = useRef<EventSource | null>(null)
-  const retryTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const mountedRef = useRef(true)
-
-  const connect = useCallback(() => {
-    // Cancel any pending reconnect timer
-    if (retryTimerRef.current) {
-      clearTimeout(retryTimerRef.current)
-      retryTimerRef.current = null
-    }
-    // Close any existing connection cleanly
-    if (eventSourceRef.current) {
-      eventSourceRef.current.close()
-      eventSourceRef.current = null
-    }
-
-    const url = `/ssf/events/stream?session=${encodeURIComponent(sessionId)}`
-    const es = new EventSource(url)
-    eventSourceRef.current = es
-
-    es.addEventListener('connected', () => {
-      setIsConnected(true)
-    })
-
-    // Pipeline events are delivered reliably through the API response
-    // (ingestResponseEvents). SSE is kept only for connection status.
-    // No 'pipeline' listener — avoids duplicates from the same broadcast
-    // reaching both the action handler's temp channels and the SSE channels.
-
-    es.onerror = () => {
-      // CRITICAL: Close immediately to prevent the browser's built-in
-      // EventSource auto-reconnect from creating duplicate backend
-      // goroutines that consume broadcast events into dead connections.
-      es.close()
-      eventSourceRef.current = null
-      setIsConnected(false)
-
-      // Reconnect on our own schedule (fixed 2s delay).
-      // The backend cancels stale goroutines per-session, so even if
-      // two reconnects overlap, only the latest survives.
-      if (mountedRef.current) {
-        retryTimerRef.current = setTimeout(() => {
-          if (mountedRef.current) connect()
-        }, 2000)
-      }
-    }
-
-    es.onopen = () => {
-      setIsConnected(true)
-    }
-  }, [sessionId])
-
-  const disconnect = useCallback(() => {
-    if (retryTimerRef.current) {
-      clearTimeout(retryTimerRef.current)
-      retryTimerRef.current = null
-    }
-    if (eventSourceRef.current) {
-      eventSourceRef.current.close()
-      eventSourceRef.current = null
-    }
-    setIsConnected(false)
-  }, [])
-
-  const clearEvents = useCallback(() => {
-    setPipelineEvents([])
-    setHttpExchanges([])
-  }, [])
-
-  // Process pipeline events returned directly in the action API response.
-  // This is the RELIABLE path — events come through the HTTP response,
-  // not the fragile SSE stream through the Vite proxy.
-  const ingestResponseEvents = useCallback((events: SSEPipelineEvent[]) => {
-    const flowEvents: FlowEvent[] = []
-    const exchanges: CapturedHTTPExchange[] = []
-
-    for (const sse of events) {
-      if (sse.event.type === 'http_exchange' && sse.event.data) {
-        exchanges.push(sse.event.data as unknown as CapturedHTTPExchange)
-        continue
-      }
-      const fe = mapSSEToFlowEvent(sse)
-      if (fe) flowEvents.push(fe)
-    }
-
-    if (flowEvents.length > 0) setPipelineEvents(prev => [...prev, ...flowEvents])
-    if (exchanges.length > 0) setHttpExchanges(prev => [...prev, ...exchanges])
-  }, [])
-
-  useEffect(() => {
-    mountedRef.current = true
-    connect()
-    return () => {
-      mountedRef.current = false
-      disconnect()
-    }
-  }, [connect, disconnect])
-
-  return { pipelineEvents, httpExchanges, isConnected, clearEvents, ingestResponseEvents, connect, disconnect }
-}
-
 // ============================================================================
 // Event Definitions
 // ============================================================================
 
-interface EventDef {
-  id: string
-  name: string
-  icon: React.ElementType
-  description: string
-  category: 'CAEP' | 'RISC'
-  rfcReference: string
-}
-
 // Icon mapping for event types
-const EVENT_ICONS: Record<string, React.ElementType> = {
+const EVENT_ICONS: Record<string, ElementType> = {
   'session-revoked': Lock,
   'token-claims-change': Code,
   'credential-change': Key,
@@ -405,7 +73,7 @@ export function SSFSandbox() {
   const [manualSET, setManualSET] = useState('')
 
   // Session ID for this browser tab
-  const sessionId = useMemo(() => getOrCreateSessionId(), [])
+  const sessionId = useMemo(() => getOrCreateSSFSessionId(), [])
 
   // Real-time SSE event stream from the backend
   const { pipelineEvents, httpExchanges, isConnected, clearEvents, ingestResponseEvents } = useSSFEventStream(sessionId)
@@ -587,7 +255,7 @@ export function SSFSandbox() {
               <span className="truncate">SSF Sandbox</span>
             </h1>
             <div className="flex items-center gap-2">
-              {status !== 'idle' && <StatusBadge status={status} />}
+              {status !== 'idle' && <SharedStatusBadge status={status} variants={SSF_STATUS_BADGE_VARIANTS} />}
               <span className="text-xs text-surface-500 font-mono hidden sm:block" title={`Session: ${sessionId}`}>
                 🔒 private session
               </span>
@@ -858,24 +526,6 @@ export function SSFSandbox() {
 // Sub-Components
 // ============================================================================
 
-function StatusBadge({ status }: { status: string }) {
-  const config: Record<string, { bg: string; border: string; text: string; label: string; shortLabel: string }> = {
-    completed: { bg: 'bg-green-500/10', border: 'border-green-500/30', text: 'text-green-400', label: 'Completed', shortLabel: 'Done' },
-    executing: { bg: 'bg-amber-500/10', border: 'border-amber-500/30', text: 'text-amber-400', label: 'Executing...', shortLabel: 'Running' },
-    error: { bg: 'bg-red-500/10', border: 'border-red-500/30', text: 'text-red-400', label: 'Error', shortLabel: 'Error' },
-  }
-  const c = config[status] || config.error
-
-  return (
-    <div className={`px-2 sm:px-3 py-1 sm:py-1.5 rounded-full ${c.bg} border ${c.border} flex-shrink-0`}>
-      <span className={`text-xs sm:text-sm font-medium ${c.text} whitespace-nowrap`}>
-        <span className="hidden sm:inline">{c.label}</span>
-        <span className="sm:hidden">{c.shortLabel}</span>
-      </span>
-    </div>
-  )
-}
-
 function QuickButton({ 
   icon: Icon, 
   label, 
@@ -909,598 +559,6 @@ function QuickButton({
         <div className="text-[10px] sm:text-sm text-surface-400">{sublabel}</div>
       </div>
       <ChevronRight className="w-4 h-4 sm:w-5 sm:h-5 text-surface-600 group-hover:text-surface-400 transition-colors flex-shrink-0" />
-    </button>
-  )
-}
-
-// ============================================================================
-// SSF Flow Panel (matches RealFlowPanel structure)
-// ============================================================================
-
-interface SSFFlowPanelProps {
-  status: 'idle' | 'executing' | 'completed' | 'error'
-  events: FlowEvent[]
-  httpExchanges: CapturedHTTPExchange[]
-  decodedSET: DecodedSET | null
-  securityState: SecurityState | null
-  activeTab: 'events' | 'traffic' | 'set' | 'state'
-  onTabChange: (tab: 'events' | 'traffic' | 'set' | 'state') => void
-  onResetState: () => void
-  selectedEvent: EventDef
-  isSSEConnected: boolean
-}
-
-function SSFFlowPanel({ 
-  status, 
-  events,
-  httpExchanges,
-  decodedSET, 
-  securityState,
-  activeTab, 
-  onTabChange,
-  onResetState,
-  selectedEvent,
-  isSSEConnected
-}: SSFFlowPanelProps) {
-  const statusConfig = {
-    idle: { icon: Play, color: 'text-surface-400', bg: 'bg-surface-800', label: 'Ready to Execute' },
-    executing: { icon: Clock, color: 'text-cyan-400', bg: 'bg-cyan-500/10', label: 'Executing...' },
-    completed: { icon: CheckCircle, color: 'text-green-400', bg: 'bg-green-500/10', label: 'Completed' },
-    error: { icon: XCircle, color: 'text-red-400', bg: 'bg-red-500/10', label: 'Error' },
-  }
-
-  const currentStatus = statusConfig[status]
-  const StatusIcon = currentStatus.icon
-
-  return (
-    <div className="space-y-3 sm:space-y-4">
-      {/* Flow Info Header */}
-      <div className="p-3 sm:p-4 rounded-xl bg-surface-900/50 border border-white/5">
-        <div className="flex items-start justify-between gap-2 mb-3">
-          <div className="flex items-center gap-2 sm:gap-3 min-w-0 flex-1">
-            <div className={`p-1.5 sm:p-2 rounded-lg flex-shrink-0 ${currentStatus.bg}`}>
-              <StatusIcon className={`w-4 h-4 ${currentStatus.color}`} />
-            </div>
-            <div className="min-w-0">
-              <p className={`text-xs sm:text-sm font-medium ${currentStatus.color}`}>{currentStatus.label}</p>
-            </div>
-          </div>
-        </div>
-        
-        <h2 className="font-medium text-white text-xs sm:text-sm leading-relaxed mb-2">{selectedEvent.description}</h2>
-        
-        <div className="flex flex-wrap items-center gap-1.5 sm:gap-2">
-          <div className="flex items-center gap-1">
-            <Book className="w-3 h-3 sm:w-4 sm:h-4 text-blue-400 flex-shrink-0" />
-            <span className="text-[10px] sm:text-xs text-blue-400 font-mono">{selectedEvent.rfcReference}</span>
-          </div>
-          <span className={`px-1.5 sm:px-2 py-0.5 rounded text-[10px] sm:text-xs ${
-            selectedEvent.category === 'CAEP' 
-              ? 'bg-blue-500/10 text-blue-400' 
-              : 'bg-amber-500/10 text-amber-400'
-          }`}>
-            {selectedEvent.category}
-          </span>
-        </div>
-      </div>
-
-      {/* SSE Connection Status */}
-      <div className="flex items-center gap-2 text-xs">
-        <div className={`w-2 h-2 rounded-full ${isSSEConnected ? 'bg-green-400 animate-pulse' : 'bg-surface-600'}`} />
-        <span className="text-surface-500">{isSSEConnected ? 'Live stream connected' : 'Connecting...'}</span>
-      </div>
-
-      {/* Tab Navigation */}
-      <div className="flex gap-1 p-1 rounded-lg bg-surface-900/50 overflow-x-auto scrollbar-hide">
-        {[
-          { id: 'events', label: 'Pipeline', count: events.length, icon: Zap },
-          { id: 'traffic', label: 'Traffic', count: httpExchanges.length, icon: Globe },
-          { id: 'set', label: 'SET', count: decodedSET ? 1 : 0, icon: Key },
-          { id: 'state', label: 'State', count: securityState ? 1 : 0, icon: User },
-        ].map(tab => (
-          <button
-            key={tab.id}
-            onClick={() => onTabChange(tab.id as typeof activeTab)}
-            className={`flex-1 min-w-0 flex items-center justify-center gap-1.5 sm:gap-2 px-2 sm:px-4 py-2.5 sm:py-2 rounded-md text-sm font-medium transition-colors whitespace-nowrap ${
-              activeTab === tab.id
-                ? 'bg-surface-800 text-white'
-                : 'text-surface-400 hover:text-white active:text-white'
-            }`}
-          >
-            <tab.icon className="w-4 h-4 flex-shrink-0" />
-            <span className="hidden sm:inline">{tab.label}</span>
-            {tab.count > 0 && (
-              <span className="px-1.5 py-0.5 rounded text-xs bg-surface-700 flex-shrink-0">
-                {tab.count}
-              </span>
-            )}
-          </button>
-        ))}
-      </div>
-
-      {/* Tab Content */}
-      <div className="min-h-[300px] sm:min-h-[400px] max-h-[450px] sm:max-h-[600px] overflow-y-auto">
-        <AnimatePresence mode="wait">
-          {activeTab === 'events' && (
-            <motion.div
-              key="events"
-              initial={{ opacity: 0, y: 10 }}
-              animate={{ opacity: 1, y: 0 }}
-              exit={{ opacity: 0, y: -10 }}
-            >
-              <EventsList events={events} />
-            </motion.div>
-          )}
-          {activeTab === 'traffic' && (
-            <motion.div
-              key="traffic"
-              initial={{ opacity: 0, y: 10 }}
-              animate={{ opacity: 1, y: 0 }}
-              exit={{ opacity: 0, y: -10 }}
-            >
-              <TrafficPanel exchanges={httpExchanges} />
-            </motion.div>
-          )}
-          {activeTab === 'set' && (
-            <motion.div
-              key="set"
-              initial={{ opacity: 0, y: 10 }}
-              animate={{ opacity: 1, y: 0 }}
-              exit={{ opacity: 0, y: -10 }}
-            >
-              <SETInspector decodedSET={decodedSET} />
-            </motion.div>
-          )}
-          {activeTab === 'state' && (
-            <motion.div
-              key="state"
-              initial={{ opacity: 0, y: 10 }}
-              animate={{ opacity: 1, y: 0 }}
-              exit={{ opacity: 0, y: -10 }}
-            >
-              <SecurityStatePanel state={securityState} onReset={onResetState} />
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </div>
-    </div>
-  )
-}
-
-// ============================================================================
-// Events List (matches Looking Glass EventsList)
-// ============================================================================
-
-function EventsList({ events }: { events: FlowEvent[] }) {
-  if (events.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center py-12 text-center">
-        <Info className="w-12 h-12 text-surface-600 mb-3" />
-        <p className="text-surface-400">No events yet</p>
-        <p className="text-surface-400 text-sm">Execute the flow to see SSF protocol events</p>
-      </div>
-    )
-  }
-
-  const eventConfig: Record<FlowEvent['type'], { icon: React.ElementType; color: string }> = {
-    info: { icon: Info, color: 'text-blue-400 bg-blue-500/10' },
-    request: { icon: Send, color: 'text-cyan-400 bg-cyan-500/10' },
-    response: { icon: ArrowDownLeft, color: 'text-green-400 bg-green-500/10' },
-    token: { icon: Key, color: 'text-yellow-400 bg-yellow-500/10' },
-    crypto: { icon: Lock, color: 'text-purple-400 bg-purple-500/10' },
-    security: { icon: Shield, color: 'text-orange-400 bg-orange-500/10' },
-    action: { icon: Bell, color: 'text-pink-400 bg-pink-500/10' },
-    error: { icon: AlertTriangle, color: 'text-red-400 bg-red-500/10' },
-  }
-
-  return (
-    <div className="space-y-2">
-      {events.map((event, index) => {
-        const config = eventConfig[event.type]
-        const Icon = config.icon
-        const [textColor, bgColor] = config.color.split(' ')
-
-        return (
-          <motion.div
-            key={event.id}
-            initial={{ opacity: 0, x: -20 }}
-            animate={{ opacity: 1, x: 0 }}
-            transition={{ delay: index * 0.03 }}
-            className="p-3 rounded-lg bg-surface-900/50 border border-white/5"
-          >
-            <div className="flex items-start gap-3">
-              <div className={`p-1.5 rounded-lg ${bgColor}`}>
-                <Icon className={`w-4 h-4 ${textColor}`} />
-              </div>
-              <div className="flex-1 min-w-0">
-                <div className="flex items-center justify-between gap-2">
-                  <h3 className="font-medium text-white text-sm">{event.title}</h3>
-                  <span className="text-xs text-surface-400 shrink-0">
-                    {event.timestamp.toLocaleTimeString()}
-                  </span>
-                </div>
-                <p className="text-sm text-surface-400 mt-0.5">{event.description}</p>
-                
-                {event.rfcReference && (
-                  <div className="mt-2">
-                    <span className="inline-flex items-center gap-1 px-2 py-0.5 rounded text-xs bg-indigo-500/10 text-indigo-400 font-mono">
-                      <Book className="w-3 h-3" />
-                      {event.rfcReference}
-                    </span>
-                  </div>
-                )}
-                
-                {event.data && Object.keys(event.data).length > 0 && (
-                  <ExpandableData data={event.data} />
-                )}
-              </div>
-            </div>
-          </motion.div>
-        )
-      })}
-    </div>
-  )
-}
-
-// ============================================================================
-// SET Inspector (matches TokensList pattern)
-// ============================================================================
-
-function SETInspector({ decodedSET }: { decodedSET: DecodedSET | null }) {
-  if (!decodedSET) {
-    return (
-      <div className="flex flex-col items-center justify-center py-12 text-center">
-        <Key className="w-12 h-12 text-surface-600 mb-3" />
-        <p className="text-surface-400">No SET captured</p>
-        <p className="text-surface-400 text-sm">Complete the flow to see the decoded SET</p>
-      </div>
-    )
-  }
-
-  return (
-    <div className="space-y-4">
-      <div className="rounded-lg bg-surface-900/50 border border-white/5 overflow-hidden">
-        <div className="p-2.5 sm:p-3 border-b border-white/5">
-          <div className="flex items-center justify-between gap-2">
-            <div className="flex items-center gap-1.5 sm:gap-2 min-w-0">
-              <Key className="w-4 h-4 sm:w-5 sm:h-5 flex-shrink-0 text-amber-400" />
-              <span className="font-medium text-white text-sm sm:text-base truncate">
-                Security Event Token (SET)
-              </span>
-            </div>
-            <div className="flex items-center gap-1.5 sm:gap-2 flex-shrink-0">
-              <span className="flex items-center gap-1 text-[10px] sm:text-xs text-green-400">
-                <CheckCircle className="w-3 h-3 sm:w-3.5 sm:h-3.5" />
-                <span className="hidden sm:inline">Valid</span>
-              </span>
-              <CopyButton text={decodedSET.raw_token} />
-            </div>
-          </div>
-        </div>
-
-        <div className="p-2.5 sm:p-3 space-y-2 sm:space-y-3">
-          {/* Header */}
-          <div>
-            <h3 className="text-[10px] sm:text-xs font-medium text-surface-400 mb-1">Header</h3>
-            <pre className="p-2 rounded bg-surface-950 text-[10px] sm:text-xs font-mono text-surface-300 overflow-x-auto scrollbar-hide">
-              {JSON.stringify(decodedSET.header, null, 2)}
-            </pre>
-          </div>
-
-          {/* Payload */}
-          <div>
-            <h3 className="text-[10px] sm:text-xs font-medium text-surface-400 mb-1">Payload</h3>
-            <pre className="p-2 rounded bg-surface-950 text-[10px] sm:text-xs font-mono text-surface-300 overflow-x-auto scrollbar-hide">
-              {JSON.stringify({
-                jti: decodedSET.jti,
-                iss: decodedSET.iss,
-                aud: decodedSET.aud,
-                iat: decodedSET.iat,
-                sub_id: decodedSET.sub_id,
-                events: decodedSET.events.map(e => ({
-                  [e.type]: e.payload
-                }))
-              }, null, 2)}
-            </pre>
-          </div>
-
-          {/* Event Metadata */}
-          {decodedSET.events[0] && (
-            <div className={`p-3 rounded-lg ${
-              decodedSET.events[0].metadata.category === 'CAEP' 
-                ? 'bg-blue-500/5 border border-blue-500/20' 
-                : 'bg-amber-500/5 border border-amber-500/20'
-            }`}>
-              <div className="flex items-center gap-2 mb-2">
-                <span className={`text-xs font-bold px-2 py-0.5 rounded ${
-                  decodedSET.events[0].metadata.category === 'CAEP' ? 'bg-blue-500 text-white' : 'bg-amber-500 text-white'
-                }`}>
-                  {decodedSET.events[0].metadata.category}
-                </span>
-                <span className="text-sm text-surface-200">{decodedSET.events[0].metadata.name}</span>
-              </div>
-              <p className="text-xs text-surface-400">{decodedSET.events[0].metadata.zero_trust_impact}</p>
-            </div>
-          )}
-
-          {/* Raw Token */}
-          <div>
-            <div className="flex items-center justify-between mb-1">
-              <h3 className="text-[10px] sm:text-xs font-medium text-surface-400">Raw Token</h3>
-              <CopyButton text={decodedSET.raw_token} />
-            </div>
-            <div className="p-2 rounded bg-surface-950 font-mono text-[10px] break-all">
-              <span className="text-red-400">{decodedSET.raw_token.split('.')[0]}</span>
-              <span className="text-surface-600">.</span>
-              <span className="text-purple-400">{decodedSET.raw_token.split('.')[1]}</span>
-              <span className="text-surface-600">.</span>
-              <span className="text-cyan-400">{decodedSET.raw_token.split('.')[2]}</span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// ============================================================================
-// Security State Panel
-// ============================================================================
-
-function SecurityStatePanel({ state, onReset }: { state: SecurityState | null; onReset?: () => void }) {
-  if (!state) {
-    return (
-      <div className="flex flex-col items-center justify-center py-12 text-center">
-        <User className="w-12 h-12 text-surface-600 mb-3" />
-        <p className="text-surface-400">No security state</p>
-        <p className="text-surface-400 text-sm">Execute a flow to see state changes</p>
-      </div>
-    )
-  }
-
-  return (
-    <div className="space-y-4">
-      {/* State Card */}
-      <div className="rounded-lg bg-surface-900/50 border border-white/5 overflow-hidden">
-        <div className="p-3 border-b border-white/5 flex items-center justify-between">
-          <div className="flex items-center gap-2">
-            <User className="w-4 h-4 text-purple-400" />
-            <span className="font-medium text-white text-sm">Security State: {state.email}</span>
-          </div>
-          {onReset && (
-            <button
-              onClick={onReset}
-              className="text-xs px-2 py-1 rounded bg-surface-800 text-surface-400 hover:text-white transition-colors"
-            >
-              Reset State
-            </button>
-          )}
-        </div>
-        <div className="p-3 grid grid-cols-2 gap-3">
-          <StateItem label="Active Sessions" value={state.sessions_active.toString()} status={state.sessions_active > 0 ? 'good' : 'neutral'} />
-          <StateItem label="Account Status" value={state.account_enabled ? 'Enabled' : 'Disabled'} status={state.account_enabled ? 'good' : 'bad'} />
-          <StateItem label="Tokens" value={state.tokens_valid ? 'Valid' : 'Invalid'} status={state.tokens_valid ? 'good' : 'bad'} />
-          <StateItem label="Password Reset" value={state.password_reset_required ? 'Required' : 'Not Required'} status={state.password_reset_required ? 'warn' : 'good'} />
-        </div>
-        <div className="px-3 pb-3 text-xs text-surface-400">
-          Last modified: {new Date(state.last_modified).toLocaleString()} by {state.modified_by}
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function StateItem({ label, value, status }: { label: string; value: string; status: 'good' | 'warn' | 'bad' | 'neutral' }) {
-  const colors = {
-    good: 'text-green-400',
-    warn: 'text-amber-400',
-    bad: 'text-red-400',
-    neutral: 'text-surface-300',
-  }
-  return (
-    <div className="p-2 rounded bg-surface-800/50">
-      <div className="text-[10px] text-surface-400 uppercase tracking-wider">{label}</div>
-      <div className={`text-sm font-medium ${colors[status]}`}>{value}</div>
-    </div>
-  )
-}
-
-// ============================================================================
-// Traffic Panel - Real HTTP Exchanges
-// ============================================================================
-
-function TrafficPanel({ exchanges }: { exchanges: CapturedHTTPExchange[] }) {
-  if (exchanges.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center py-12 text-center">
-        <Globe className="w-12 h-12 text-surface-600 mb-3" />
-        <p className="text-surface-400">No HTTP traffic captured</p>
-        <p className="text-surface-500 text-sm mt-1">Execute the flow to see real HTTP exchanges<br />between the transmitter and receiver</p>
-      </div>
-    )
-  }
-
-  return (
-    <div className="space-y-2">
-      {exchanges.map((exchange, index) => (
-        <HTTPExchangeCard key={index} exchange={exchange} />
-      ))}
-    </div>
-  )
-}
-
-/** Matches the Looking Glass ExchangeCard pattern exactly */
-function HTTPExchangeCard({ exchange }: { exchange: CapturedHTTPExchange }) {
-  const [isExpanded, setIsExpanded] = useState(false)
-
-  const statusCode = exchange.response.status_code
-  const statusBg = statusCode >= 200 && statusCode < 300
-    ? 'bg-green-500/10 text-green-400'
-    : statusCode >= 400
-    ? 'bg-red-500/10 text-red-400'
-    : 'bg-yellow-500/10 text-yellow-400'
-  const statusTextColor = statusCode >= 200 && statusCode < 300
-    ? 'text-green-400' : statusCode >= 400 ? 'text-red-400' : 'text-yellow-400'
-
-  // Extract RFC from label (e.g. "Push Delivery (RFC 8935)" -> "RFC 8935")
-  const rfcMatch = exchange.label.match(/\(([^)]+)\)/)
-  const rfcReference = rfcMatch ? rfcMatch[1] : null
-  const stepLabel = exchange.label.replace(/\s*\([^)]*\)/, '')
-
-  return (
-    <div className="rounded-lg bg-surface-900/50 border border-white/5 overflow-hidden">
-      <button
-        onClick={() => setIsExpanded(!isExpanded)}
-        className="w-full p-3 flex items-center gap-2 sm:gap-3 hover:bg-white/5 active:bg-white/10 transition-colors text-left"
-      >
-        <div className="p-1.5 rounded bg-cyan-500/10 flex-shrink-0">
-          <Send className="w-4 h-4 text-cyan-400" />
-        </div>
-        <div className="flex-1 min-w-0 overflow-hidden">
-          <div className="flex items-center gap-2 flex-wrap sm:flex-nowrap">
-            <span className="font-mono text-xs sm:text-sm font-medium text-cyan-400 flex-shrink-0">
-              {exchange.request.method}
-            </span>
-            <span className="text-xs sm:text-sm text-surface-300 truncate max-w-full">
-              {exchange.request.url}
-            </span>
-          </div>
-          <p className="text-xs text-surface-400 mt-0.5 truncate">{stepLabel}</p>
-        </div>
-        <div className="flex items-center gap-1.5 sm:gap-2 flex-shrink-0">
-          <span className={`px-1.5 sm:px-2 py-0.5 rounded text-xs font-medium ${statusBg}`}>
-            {statusCode}
-          </span>
-          {rfcReference && (
-            <span className="hidden sm:inline px-1.5 py-0.5 rounded text-xs bg-indigo-500/10 text-indigo-400 font-mono">
-              {rfcReference}
-            </span>
-          )}
-          {isExpanded ? (
-            <ChevronDown className="w-4 h-4 text-surface-400" />
-          ) : (
-            <ChevronRight className="w-4 h-4 text-surface-400" />
-          )}
-        </div>
-      </button>
-
-      <AnimatePresence>
-        {isExpanded && (
-          <motion.div
-            initial={{ height: 0, opacity: 0 }}
-            animate={{ height: 'auto', opacity: 1 }}
-            exit={{ height: 0, opacity: 0 }}
-            className="border-t border-white/5"
-          >
-            <div className="p-3 space-y-3">
-              {/* Request */}
-              <div>
-                <h4 className="text-xs font-medium text-surface-400 mb-2">Request</h4>
-                <pre className="p-2 rounded bg-surface-950 text-xs font-mono overflow-x-auto">
-                  <div className="text-cyan-400">
-                    {exchange.request.method} {exchange.request.url}
-                  </div>
-                  {exchange.request.headers && Object.entries(exchange.request.headers).map(([k, v]) => (
-                    <div key={k} className="text-surface-400">
-                      <span className="text-surface-400">{k}:</span> {v}
-                    </div>
-                  ))}
-                  {exchange.request.body && (
-                    <div className="mt-2 pt-2 border-t border-white/5 text-surface-300">
-                      {tryFormatJSON(exchange.request.body)}
-                    </div>
-                  )}
-                </pre>
-              </div>
-
-              {/* Response */}
-              <div>
-                <h4 className="text-xs font-medium text-surface-400 mb-2">
-                  Response ({exchange.duration_ms}ms)
-                </h4>
-                <pre className="p-2 rounded bg-surface-950 text-xs font-mono overflow-x-auto">
-                  <div className={statusTextColor}>
-                    {statusCode} {statusCode === 202 ? 'OK' : statusCode < 300 ? 'OK' : statusCode < 400 ? 'Redirect' : 'Error'}
-                  </div>
-                  {exchange.response.body && (
-                    <div className="mt-2 text-surface-300">
-                      {tryFormatJSON(exchange.response.body)}
-                    </div>
-                  )}
-                </pre>
-              </div>
-            </div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
-  )
-}
-
-function tryFormatJSON(str: string): string {
-  try {
-    return JSON.stringify(JSON.parse(str), null, 2)
-  } catch {
-    return str
-  }
-}
-
-// ============================================================================
-// Utility Components
-// ============================================================================
-
-function ExpandableData({ data }: { data: Record<string, unknown> }) {
-  const [isExpanded, setIsExpanded] = useState(false)
-  
-  if (Object.keys(data).length === 0) return null
-
-  return (
-    <div className="mt-2">
-      <button
-        onClick={() => setIsExpanded(!isExpanded)}
-        className="flex items-center gap-1 text-xs text-surface-400 hover:text-surface-300 transition-colors"
-      >
-        <Code className="w-3 h-3" />
-        {isExpanded ? 'Hide' : 'Show'} data
-        {isExpanded ? <ChevronDown className="w-3 h-3" /> : <ChevronRight className="w-3 h-3" />}
-      </button>
-      
-      <AnimatePresence>
-        {isExpanded && (
-          <motion.pre
-            initial={{ height: 0, opacity: 0 }}
-            animate={{ height: 'auto', opacity: 1 }}
-            exit={{ height: 0, opacity: 0 }}
-            className="mt-2 p-2 rounded bg-surface-950 text-xs font-mono text-surface-400 overflow-x-auto"
-          >
-            {JSON.stringify(data, null, 2)}
-          </motion.pre>
-        )}
-      </AnimatePresence>
-    </div>
-  )
-}
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false)
-
-  const handleCopy = async () => {
-    await navigator.clipboard.writeText(text)
-    setCopied(true)
-    setTimeout(() => setCopied(false), 2000)
-  }
-
-  return (
-    <button
-      onClick={handleCopy}
-      className="p-1.5 sm:p-1 rounded hover:bg-white/10 active:bg-white/20 transition-colors flex-shrink-0"
-      title="Copy to clipboard"
-    >
-      {copied ? (
-        <Check className="w-3.5 h-3.5 sm:w-3 sm:h-3 text-green-400" />
-      ) : (
-        <Copy className="w-3.5 h-3.5 sm:w-3 sm:h-3 text-surface-400" />
-      )}
     </button>
   )
 }
